+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 #### Upcoming
 
+*Minor*
+
+* Changed `choose` command for Scapegoat to take a space separated list rather than comma separated. ([#98](https://github.com/hjwylde/werewolf/issues/98))
+
 #### v0.4.8.0
 
 *Minor*